usw switch statement to get parser
