--- Annoy-o-Bot/JsonReminderParser.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using Newtonsoft.Json;
-
-namespace Annoy_o_Bot
-{
-    public class JsonReminderParser
-    {
-        public static Reminder Parse(string documentContent)
-        {
-            var reminder = JsonConvert.DeserializeObject<Reminder>(documentContent);
-
-            if (string.IsNullOrWhiteSpace(reminder.Title))
-            {
-                throw new ArgumentException("A reminder must provide a non-empty Title property");
-            }
-            return reminder;
-        }
-    }
-}
